new rendering of cumulative filter (#1)
* render categories

* new rendering of cumulative filter
