+++ src/Resources/contao/languages/en/modules.php.bak
@@ -13,4 +13,3 @@
  */
 $GLOBALS['TL_LANG']['FMD']['newscategories'] = ['News categories list', 'Adds a categories list to the page.'];
 $GLOBALS['TL_LANG']['FMD']['newscategories_cumulative'] = ['News categories cumulative filter', 'Adds a cumulative filter of categories to the page.'];
-$GLOBALS['TL_LANG']['FMD']['newscategories_cumulativehierarchical'] = ['News categories cumulative hierarchical filter', 'Adds a cumulative hierarchical filter of categories to the page.'];
