+++ src/Resources/contao/templates/modules/mod_newscategories_cumulativehierarchical.html5
@@ -1,0 +1,7 @@
+<?php $this->extend('block_unsearchable'); ?>
+
+<?php $this->block('content'); ?>
+
+<?= $this->categories ?>
+
+<?php $this->endblock(); ?>
